Correct Licencing Information.
This may seem a bit grotesque, even illegal, but it's actually not.
The copyright holder, is the writer of the code, the code is not releaseable without his concent, and his agreement on the licence (Especially under french law where SmartJog S.A.S. operates).

MpegParse, was always GNU GPL, and always will be.

for the email part, there is no point releasing software with email that doen't work anymore.

The last point of quadripilorisitycutting would be the actual text modified by this patch that could be LGPL, copyrighted by SmartJog S.A.S., and modified by me. I leave prior art demonstration as an exercise to the reader.

Signed-off-by: Niv Sardi <xaiki@evilgiggle.com>